feat(next diff): receive a `showPatchTicket` event from the ide
