+++ src/components/Markdown/Markdown.tsx
@@ -1,4 +1,4 @@
-import React, { Key, useCallback, useMemo, useState } from "react";
+import React, { Key, useMemo } from "react";
 import ReactMarkdown, { Components } from "react-markdown";
 import remarkBreaks from "remark-breaks";
 import classNames from "classnames";
@@ -25,15 +25,10 @@ import {
 import rehypeKatex from "rehype-katex";
 import remarkMath from "remark-math";
 import "katex/dist/katex.min.css";
-import { diffApi, isDetailMessage } from "../../services/refact";
-import { useEventsBusForIDE } from "../../hooks";
-import { useSelector } from "react-redux";
+import { usePatchActions } from "../../hooks";
+
 import { ErrorCallout, DiffWarningCallout } from "../Callout";
-import {
-  selectIsStreaming,
-  selectIsWaiting,
-  selectMessages,
-} from "../../features/Chat";
+
 import { TruncateLeft } from "../Text";
 
 export type MarkdownProps = Pick<
@@ -45,123 +40,6 @@ export type MarkdownProps = Pick<
     MarkdownCodeBlockProps,
     "startingLineNumber" | "showLineNumbers" | "useInlineStyles" | "style"
   > & { canHavePins?: boolean };
-
-const usePinActions = () => {
-  const {
-    diffPreview,
-    startFileAnimation,
-    stopFileAnimation,
-    openFile,
-    writeResultsToFile,
-  } = useEventsBusForIDE();
-  const messages = useSelector(selectMessages);
-  const isStreaming = useSelector(selectIsStreaming);
-  const isWaiting = useSelector(selectIsWaiting);
-
-  const [errorMessage, setErrorMessage] = useState<{
-    type: "warning" | "error";
-    text: string;
-  } | null>(null);
-
-  const resetErrorMessage = useCallback(() => {
-    setErrorMessage(null);
-  }, []);
-
-  const [getPatch, patchResult] =
-    diffApi.usePatchSingleFileFromTicketMutation();
-
-  const disable = useMemo(() => {
-    return !!errorMessage || isStreaming || isWaiting || patchResult.isLoading;
-  }, [errorMessage, isStreaming, isWaiting, patchResult.isLoading]);
-
-  const handleShow = useCallback(
-    (pin: string) => {
-      const [, , fileName] = pin.split(" ");
-      startFileAnimation(fileName);
-      getPatch({ pin, messages })
-        .unwrap()
-        .then((maybeDetail) => {
-          if (isDetailMessage(maybeDetail)) {
-            const error = new Error(maybeDetail.detail);
-            throw error;
-          }
-          return maybeDetail;
-        })
-        .then((patch) => {
-          stopFileAnimation(fileName);
-          // TODO: might work with patch results?
-          diffPreview(patch);
-        })
-        .catch((error: Error | { data: { detail: string } }) => {
-          stopFileAnimation(fileName);
-          if ("message" in error) {
-            setErrorMessage({
-              type: "error",
-              text: "Failed to open patch: " + error.message,
-            });
-          } else {
-            setErrorMessage({
-              type: "error",
-              text: "Failed to open patch: " + error.data.detail,
-            });
-          }
-        });
-    },
-    [diffPreview, getPatch, messages, startFileAnimation, stopFileAnimation],
-  );
-
-  const handleApply = useCallback(
-    (pin: string) => {
-      const [, , fileName] = pin.split(" ");
-      startFileAnimation(fileName);
-
-      getPatch({ pin, messages })
-        .unwrap()
-        .then((maybeDetail) => {
-          if (isDetailMessage(maybeDetail)) {
-            const error = new Error(maybeDetail.detail);
-            throw error;
-          }
-          return maybeDetail;
-        })
-        .then((patch) => {
-          stopFileAnimation(fileName);
-          writeResultsToFile(patch.results);
-        })
-        .catch((error: Error | { data: { detail: string } }) => {
-          stopFileAnimation(fileName);
-          if ("message" in error) {
-            setErrorMessage({
-              type: "error",
-              text: "Failed to apply patch: " + error.message,
-            });
-          } else {
-            setErrorMessage({
-              type: "error",
-              text: "Failed to apply patch: " + error.data.detail,
-            });
-          }
-        });
-    },
-    [
-      getPatch,
-      messages,
-      startFileAnimation,
-      stopFileAnimation,
-      writeResultsToFile,
-    ],
-  );
-
-  return {
-    errorMessage,
-    handleShow,
-    patchResult,
-    handleApply,
-    resetErrorMessage,
-    disable,
-    openFile,
-  };
-};
 
 const MaybePinButton: React.FC<{
   key?: Key | null;
@@ -176,7 +54,7 @@ const MaybePinButton: React.FC<{
     resetErrorMessage,
     disable,
     openFile,
-  } = usePinActions();
+  } = usePatchActions();
 
   if (isPin && children.startsWith("📍OTHER")) {
     return null;
