+++ src/hooks/index.ts
@@ -20,3 +20,4 @@ export * from "./useConfig";
 export * from "./useAppDispatch";
 export * from "./useAppSelector";
 export * from "./useSendChatRequest";
+export * from "./usePatchActions";
